Several data sets for the 'tdcm' package.

Several data sets for the 'tdcm' package.

Usage

data.tdcm01

data.tdcm02

data.tdcm03

data.tdcm04

data.tdcm05

Format

data.tdcm01 is simulated sample data that has two time points, four attributes, twenty items, one group of size 1000, and a single Q-matrix. The format is a list of two:

data: a data frame of binary item responses

qmatrix: a data frame specifying the Q-matrix

data.tdcm02 is simulated data that has three time points, two attributes, ten items, one group of size 2500, and a single Q-matrix. The format is a list of two:

data: a data frame of binary item responses

qmatrix: a data frame specifying the Q-matrix

data.tdcm03 is simulated data that has three time points, two attributes, one group of size 1500, and three different ten-item Q-matrices for each time point. Anchor items are specified as items 1/1/21 and items 14/24. The format is a list of five:

data: a data frame of binary item responses

q1: a data frame specifying the Q-matrix for the first time point

q2: a data frame specifying the Q-matrix for the second time point

q3: a data frame specifying the Q-matrix for the third time point

qmatrix123: data frame specifying the combined Q-matrix for all time points

data.tdcm04 is simulated data that has two time points, four attributes, twenty items, two group of size 800 and 900, respectively, and a single Q-matrix. The format is a list of three:

data: a data frame of binary item responses

qmatrix: a data frame specifying the Q-matrix

groups: a vector specifying the examinee group memberships

data.tdcm05 is simulated data that has two has one time point, four attributes, and twenty items. For use with the 1-PLCDM. The format is a list of two:

data: a data frame of binary item responses

qmatrix: a data frame specifying the Q-matrix

Examples

## Example 1: T = 2, A = 4
data(data.tdcm01, package = "TDCM")
dat1 <- data.tdcm01$data
qmat1 <- data.tdcm01$qmatrix

# estimate TDCM with invariance assumed and full LCDM
m1 <- TDCM::tdcm(dat1, qmat1, time.points = 2, invariance = TRUE, dcmrule = "GDINA")

# summarize results with tdcm.summary function
results <- TDCM::tdcm.summary(m1, time.points = 2)
results$item.parameters
#>         λ0     λ1,1  λ1,2  λ1,3  λ1,4  λ2,12 λ2,13 λ2,14 λ2,23 λ2,24
#> Item 1  -1.923 2.616   --    --    --    --    --    --    --    -- 
#> Item 2  -2.071 2.506   --    --    --    --    --    --    --    -- 
#> Item 3  -1.936 2.506   --    --    --    --    --    --    --    -- 
#> Item 4  -1.891 1.051 1.471   --    --  1.115   --    --    --    -- 
#> Item 5  -2.157 1.705   --  1.732   --    --  0.835   --    --    -- 
#> Item 6  -1.841   --  2.175   --    --    --    --    --    --    -- 
#> Item 7  -1.841   --  2.272   --    --    --    --    --    --    -- 
#> Item 8  -1.965   --  2.475   --    --    --    --    --    --    -- 
#> Item 9  -2.029   --  1.242 1.531   --    --    --    --  1.628   -- 
#> Item 10 -2.004   --  1.921   --  1.239   --    --    --    --  0.999
#> Item 11 -1.851   --    --  2.349   --    --    --    --    --    -- 
#> Item 12 -2.045   --    --  2.566   --    --    --    --    --    -- 
#> Item 13 -2.083   --    --  2.576   --    --    --    --    --    -- 
#> Item 14 -2.125   --    --  1.739 2.104   --    --    --    --    -- 
#> Item 15 -1.805 0.777   --  1.31    --    --  1.896   --    --    -- 
#> Item 16 -2.156   --    --    --  2.736   --    --    --    --    -- 
#> Item 17 -2.089   --    --    --  2.679   --    --    --    --    -- 
#> Item 18 -2.087   --    --    --  2.476   --    --    --    --    -- 
#> Item 19 -2.11  2.219   --    --  1.46    --    --  0.558   --    -- 
#> Item 20 -2.047   --  2.408   --  1.49    --    --    --    --  0.154
#>         λ2,34
#> Item 1    -- 
#> Item 2    -- 
#> Item 3    -- 
#> Item 4    -- 
#> Item 5    -- 
#> Item 6    -- 
#> Item 7    -- 
#> Item 8    -- 
#> Item 9    -- 
#> Item 10   -- 
#> Item 11   -- 
#> Item 12   -- 
#> Item 13   -- 
#> Item 14 0.493
#> Item 15   -- 
#> Item 16   -- 
#> Item 17   -- 
#> Item 18   -- 
#> Item 19   -- 
#> Item 20   -- 
results$growth
#>             T1[1] T2[1]
#> Attribute 1 0.201 0.372
#> Attribute 2 0.327 0.492
#> Attribute 3 0.397 0.573
#> Attribute 4 0.252 0.696
results$transition.probabilities
#> , , Attribute 1: Time 1 to Time 2
#> 
#>        T2 [0] T2 [1]
#> T1 [0]  0.678  0.322
#> T1 [1]  0.432  0.568
#> 
#> , , Attribute 2: Time 1 to Time 2
#> 
#>        T2 [0] T2 [1]
#> T1 [0]  0.578  0.422
#> T1 [1]  0.363  0.637
#> 
#> , , Attribute 3: Time 1 to Time 2
#> 
#>        T2 [0] T2 [1]
#> T1 [0]   0.55   0.45
#> T1 [1]   0.24   0.76
#> 
#> , , Attribute 4: Time 1 to Time 2
#> 
#>        T2 [0] T2 [1]
#> T1 [0]  0.365  0.635
#> T1 [1]  0.123  0.877
#> 

# estimate TDCM with invariance assumed and only main effects
m2 <- TDCM::tdcm(dat1, qmat1, time.points = 2, invariance = TRUE, dcmrule = "GDINA1")

# estimate TDCM with invariance not assumed
m3 <- TDCM::tdcm(dat1, qmat1, time.points = 2, invariance = FALSE, dcmrule = "GDINA")

# compare models to assess measurement invariance.
TDCM::tdcm.compare(m1, m3)
#>   Model   loglike Deviance Npars      AIC      BIC Chisq df      p
#> 1    m1 -21429.51 42859.02    93 43045.02 43501.44 58.14 56 0.3965
#> 2    m3 -21400.44 42800.88   149 43098.88 43830.14    NA NA     NA

## Example 2: T = 3, A = 2
data(data.tdcm02, package = "TDCM")
dat2 <- data.tdcm02$data
qmat2 <- data.tdcm02$qmatrix

# estimate TDCM with invariance assumed and full LCDM
m1 <- TDCM::tdcm(dat2, qmat2, time.points = 3)

# estimate TDCM with invariance not assumed
m2 <- TDCM::tdcm(dat2, qmat2, time.points = 3, invariance = FALSE)

# compare models to assess measurement invariance
TDCM::tdcm.compare(m1, m2)
#>   Model   loglike Deviance Npars      AIC      BIC Chisq df      p
#> 1    m1 -46110.97 92221.95    50 92321.95 92613.15  73.8 56 0.0556
#> 2    m2 -46074.07 92148.13   106 92360.13 92977.48    NA NA     NA