Fitting Transition Diagnostic Classification Models (TDCMs)

The tdcm Function calibrates and fits the transition diagnostic classification model (TDCM; Madison & Bradshaw, 2018a), a longitudinal extension of the log-linear cognitive diagnosis model (LCDM; Henson, Templin, & Willse, 2009). To calibrate a TDCM involving multiple groups, refer to the mgtdcm function.

Usage

tdcm(ir_data, q_data, invariance = TRUE, dcm_rule = "GDINA")

Arguments

ir_data

The item response data, a required \(T \times N \times I\) array where \(T\) is the number of time points, \(N\) is the number of responses, and \(I\) is the number of items. The ir_from() function can be used to help prepare this argument using values from a two-dimensional array, data.frame, or matrix.

q_data

The Q-matrix data, a required \(T \times I \times A\) array where \(T\) is the number of time points, \(I\) is the number of items, and \(A\) is the number of attributes. The qmatrix_from() function can be used to help prepare this argument using values from a two-dimensional array, data.frame, or matrix.

invariance

A logical (i.e., boolean) value that indicates whether item parameter invariance is constrained to be equal at each time point. The default value is TRUE. If set to FALSE, item parameters are not assumed to be equal over time.

dcm_rule

A string or vector of condensation rules. Refer to the rule parameter of CDM::gdina for more information. The default value is "GDINA". If dcm_rule is a string, the condensation rule applies to all items. If dcm_rule is a vector, then the condesnation rules are applied item-wise.

Value

An object of class tdcm with entries as indicated in the CDM package. For the TDCM-specific results (e.g., growth, transitions), results are summarized using the tdcm.summary() function.

Details

Estimation is performed using the CDM package according to the approach described in Madison et al. (2023).

References

de la Torre, J. (2011). The generalized DINA model framework. Psychometrika, 76, 179-199.

George, A. C., Robitzsch, A., Kiefer, T., Gross, J., & Ünlü , A. (2016). The R package CDM for cognitive diagnosis models. Journal of Statistical Software, 74(2), 1-24.

Henson, R., Templin, J., & Willse, J. (2009). Defining a family of cognitive diagnosis models using log linear models with latent variables. Psychometrika, 74, 191-21.

Madison, M. J., & Bradshaw, L. (2018a). Assessing growth in a diagnostic classification model framework. Psychometrika, 82(4), 963-990.

Madison, M. J., & Bradshaw, L. (2018b). Evaluating intervention effects in a diagnostic classification model framework. Journal of Educational Measurement, 55(1), 32-51.

Madison, M.J., Chung, S., Kim, J., & Bradshaw, L. (2023). Approaches to estimating longitudinal diagnostic classification models. Behaviormetrika.

Rupp, A. A., Templin, J., & Henson, R. (2010). Diagnostic measurement: Theory, methods, and applications. New York: Guilford.