Estimating the Transition Diagnostic Classification Model (TDCM)

tdcm() is used to estimate the transition diagnostic classification model (TDCM; Madison & Bradshaw, 2018a), which is a longitudinal extension of the log-linear cognitive diagnosis model (LCDM; Henson, Templin, & Willse, 2009). For the multigroup TDCM, see mg.tdcm(). It allows for the specification of many specific DCMs via the rule option. The default DCM rule and link function specifies the LCDM. The rule can be changed to estimate the DINA model, DINO model, CRUM (i.e., ACDM, or main effects model), or reduced interaction versions of the LCDM. The link function can be changed to specify the GDINA model.

Usage

tdcm(
  data,
  q.matrix,
  num.time.points,
  invariance = TRUE,
  rule = "LCDM",
  linkfct = "logit",
  num.q.matrix = 1,
  num.items = c(),
  anchor = c(),
  forget.att = c(),
  progress = TRUE
)

Arguments

data

A required \(N \times T \times I\) data matrix containing binary item responses. For each time point, the binary item responses are in the columns.

q.matrix

A required \(I \times A\) matrix indicating which items measure which attributes. If there are multiple Q-matrices, then they must have the same number of attributes and must be stacked on top of each other for estimation (to specify multiple Q-matrices, see num.q.matrix, num.items, and anchor).

num.time.points

A required integer \(\ge 2\) specifying the number of time points (i.e., measurement occasions).

invariance

logical. If TRUE (the default), then item parameters will be constrained to be equal at each time point. If FALSE, item parameters are not assumed to be equal over time.

rule

A string or a vector indicating the specific DCM to be employed. Currently accepts "LCDM", "DINA", "DINO", "CRUM", "RRUM", "LCDM1" for the LCDM with only main effects, "LCDM2" for the LCDM with two-way interactions, "LCDM3", and so on. If rule is supplied as a single string, then that DCM will be assumed for each item. If entered as a vector, a rule can be specified for each item.

linkfct

A string or a vector indicating the LCDM link function. Currently accepts "logit" (default) to estimate the LCDM. Can be specified "identity" to estimate the GDINA model. Also accepts a "log" link function.

num.q.matrix

An optional integer specifying the number of Q-matrices. For many applications, the same assessment is administered at each time point and this number is 1 (the default). If there are different Q-matrices for each time point, then this argument must be specified and should be equal to the number of time points. For example, if there are three time points, and the Q-matrix for each time point is different, then num.q.matrix = 3. If there are three time points, and the Q-matrix is different only for time point 3, then num.q.matrix is still specified as 3.

num.items

An integer specifying the number of items. When there are multiple Q-matrices, the number of items in each Q-matrix is specified as a vector. For example, if there are three time points, and the Q-matrices for each time point have 8, 10, and 12 items, respectively, then num.items = c(8, 10, 12).

anchor

When there are different tests at each time point, this optional argument is a vector of pairs of item numbers indicating which items are the same across time points and should be held invariant. For example, if there are three Q-matrices with 10 items each, and Items 1, 11, and 21 are the same, and Items 14 and 24 are the same, then anchor = c(1,11,1,21,14,24). Default is an empty vector to indicate absence of anchor items. Note: when anchor is specified, invariance is automatically set to false for non-anchor items.

forget.att

An optional vector allowing for constraining of individual attribute proficiency loss, or forgetting. The default allows forgetting for each measured attribute (e.g., \(P(1 \rightarrow 0) \neq 0\)). This vector is specified to indicate the attributes for which forgetting is not permitted.

progress

logical. If FALSE, the function will print the progress of estimation. If TRUE (default), no progress information is printed.

Value

An object of class gdina with entries as described in CDM::gdina(). To see a TDCM-specific summary of the object (e.g., growth, transitions), use tdcm.summary().

Details

Estimation of the TDCM via the CDM package (George, et al., 2016), which is based on an EM algorithm as described in de la Torre (2011). The estimation approach is further detailed in Madison et al. (2023).

References

de la Torre, J. (2011). The Generalized DINA Model Framework. Psychometrika 76, 179–199. doi:10.1007/s11336-011-9207-7

George, A. C., Robitzsch, A., Kiefer, T., Gross, J., & Ünlü , A. (2016). The R package CDM for Cognitive Diagnosis Models. Journal of Statistical Software, 74(2), 1-24. doi:10.18637/jss.v074.i02

Henson, R., Templin, J., & Willse, J. (2009). Defining a Family of Cognitive Diagnosis Models Using Log-Linear Models with Latent Variables. Psychometrika, 74, 191-21. doi:10.1007/s11336-008-9089-5

Johnson, M. S., & Sinharay, S. (2020). The Reliability of the Posterior Probability of Skill Attainment in Diagnostic Classification Models. Journal of Educational Measurement, 47(1), 5–31. doi:10.3102/1076998619864550

Kaya, Y., & Leite, W. (2017). Assessing Change in Latent Skills Across Time With Longitudinal Cognitive Diagnosis Modeling: An Evaluation of Model Performance. Educational and Psychological Measurement, 77(3), 369–388. doi:10.1177/0013164416659314

Li, F., Cohen, A., Bottge, B., & Templin, J. (2015). A Latent Transition Analysis Model for Assessing Change in Cognitive Skills. Educational and Psychological Measurement, 76(2), 181–204. doi:10.1177/0013164415588946

Madison, M. J. (2019). Reliably Assessing Growth with Longitudinal Diagnostic Classification Models. Educational Measurement: Issues and Practice, 38(2), 68-78. doi:10.1111/emip.12243

Madison, M. J., & Bradshaw, L. (2018a). Assessing Growth in a Diagnostic Classification Model Framework. Psychometrika, 83(4), 963-990. doi:10.1007/s11336-018-9638-5

Madison, M. J., & Bradshaw, L. (2018b). Evaluating Intervention Effects in a Diagnostic Classification Model Framework. Journal of Educational Measurement, 55(1), 32-51. doi:10.1111/jedm.12162

Madison, M.J., Chung, S., Kim, J., & Bradshaw, L.P. (2024) Approaches to estimating longitudinal diagnostic classification models. Behaviormetrika, 51(7), 7-19. doi:10.1007/s41237-023-00202-5

Rupp, A. A., Templin, J., & Henson, R. (2010). Diagnostic Measurement: Theory, Methods, and Applications. New York: Guilford. ISBN: 9781606235430.

Schellman, M., & Madison, M. J. (2024). Estimating the reliability of skill transition in longitudinal DCMs. Journal of Educational and Behavioral Statistics.

Templin, J., & Bradshaw, L. (2013). Measuring the Reliability of Diagnostic Classification Model Examinee Estimates. Journal of Classification, 30, 251-275. doi:10.1007/s00357-013-9129-4

Wang. S., Yang. Y., Culpepper, S. A., & Douglas, J. (2018). Tracking Skill Acquisition With Cognitive Diagnosis Models: A Higher-Order, Hidden Markov Model With Covariates. Journal of Educational and Behavioral Statistics, 43(1), 57-87. doi:10.3102/1076998617719727

Examples

# \donttest{
## Example 1: T = 2, A = 4
data(data.tdcm01, package = "TDCM")
data <- data.tdcm01$data
q.matrix <- data.tdcm01$q.matrix

# Estimate full TDCM with invariance assumed.
model1 <- TDCM::tdcm(data, q.matrix, num.time.points = 2)
#> [1] Preparing data for tdcm()...
#> [1] Estimating the multigroup TDCM in mg.tdcm()...
#> [1] Depending on model complexity, estimation time may vary...
#> [1] TDCM estimation complete.
#> [1] Use tdcm.summary() to display results.

# Summarize results with tdcm.summary().
results <- TDCM::tdcm.summary(model1)
#> [1] Summarizing results...
#> [1] Routine finished. Check results.
results$item.parameters
#>         λ0     λ1,1  λ1,2  λ1,3  λ1,4  λ2,12 λ2,13 λ2,14 λ2,23 λ2,24 
#> Item 1  -1.905 2.599   --    --    --    --    --    --    --    --  
#> Item 2  -2.072 2.536   --    --    --    --    --    --    --    --  
#> Item 3  -1.934 2.517   --    --    --    --    --    --    --    --  
#> Item 4  -1.892 1.091 1.499   --    --  1.057   --    --    --    --  
#> Item 5  -2.17  1.456   --  1.794   --    --  1.018   --    --    --  
#> Item 6  -1.843   --  2.199   --    --    --    --    --    --    --  
#> Item 7  -1.825   --  2.259   --    --    --    --    --    --    --  
#> Item 8  -1.967   --  2.497   --    --    --    --    --    --    --  
#> Item 9  -2.009   --  1.079 1.511   --    --    --    --  1.818   --  
#> Item 10 -2       --  1.849   --  1.324   --    --    --    --  1.065 
#> Item 11 -1.845   --    --  2.329   --    --    --    --    --    --  
#> Item 12 -2.033   --    --  2.539   --    --    --    --    --    --  
#> Item 13 -2.071   --    --  2.55    --    --    --    --    --    --  
#> Item 14 -2.093   --    --  1.739 2.031   --    --    --    --    --  
#> Item 15 -1.785 0.307   --  1.295   --    --  2.374   --    --    --  
#> Item 16 -2.218   --    --    --  2.837   --    --    --    --    --  
#> Item 17 -2.084   --    --    --  2.69    --    --    --    --    --  
#> Item 18 -2.101   --    --    --  2.521   --    --    --    --    --  
#> Item 19 -2.1   2.653   --    --  1.432   --    --  0.098   --    --  
#> Item 20 -2.061   --  2.545   --  1.53    --    --    --    --  -0.005
#>         λ2,34
#> Item 1    -- 
#> Item 2    -- 
#> Item 3    -- 
#> Item 4    -- 
#> Item 5    -- 
#> Item 6    -- 
#> Item 7    -- 
#> Item 8    -- 
#> Item 9    -- 
#> Item 10   -- 
#> Item 11   -- 
#> Item 12   -- 
#> Item 13   -- 
#> Item 14 0.496
#> Item 15   -- 
#> Item 16   -- 
#> Item 17   -- 
#> Item 18   -- 
#> Item 19   -- 
#> Item 20   -- 
results$growth
#>             T1[1] T2[1]
#> Attribute 1 0.190 0.370
#> Attribute 2 0.317 0.491
#> Attribute 3 0.392 0.579
#> Attribute 4 0.242 0.693
results$transition.probabilities
#> , , Attribute 1: Time 1 to Time 2
#> 
#>        T2 [0] T2 [1]
#> T1 [0]  0.680  0.320
#> T1 [1]  0.417  0.583
#> 
#> , , Attribute 2: Time 1 to Time 2
#> 
#>        T2 [0] T2 [1]
#> T1 [0]  0.581  0.419
#> T1 [1]  0.353  0.647
#> 
#> , , Attribute 3: Time 1 to Time 2
#> 
#>        T2 [0] T2 [1]
#> T1 [0]  0.549  0.451
#> T1 [1]  0.221  0.779
#> 
#> , , Attribute 4: Time 1 to Time 2
#> 
#>        T2 [0] T2 [1]
#> T1 [0]  0.371  0.629
#> T1 [1]  0.104  0.896
#> 
# }